Steganography Tool/requirements.txt
Flask>=2.0.3
Pillow>=8.3.1